requirements.txt
# Core ML — versions compatible with Python 3.13
lightgbm>=4.5.0
scikit-learn>=1.5.0
numpy>=2.0.0
pandas>=2.2.2
joblib>=1.4.2

# Explainability
shap>=0.45.0

# LangChain / LangGraph ecosystem
langchain>=0.3.0
langchain-core>=0.3.0
langchain-anthropic>=0.2.0
langgraph>=0.2.0
langsmith>=0.1.77

# API layer
fastapi>=0.115.0
uvicorn[standard]>=0.30.1
httpx>=0.27.0
pydantic>=2.7.4

# Frontend
streamlit>=1.35.0

# PDF generation
reportlab>=4.2.0

# Data download
kaggle>=1.6.14

# Utilities
python-dotenv>=1.0.1
pyarrow>=16.1.0

# Causal ML — install separately if C++ build tools are available
# causalml>=0.15.0
